Disable browser UI tests in CI
UI tests involving the built in browser in iOS are flaky. 
UI elements randomly don't appear. Trying to increase timeouts doesn't help.
This change disabled browser UI tests and runs only native UI tests in CI.
